runtime: switch to one runqueue per-kernel-thread
Previously, the runtime would mantain a runqueue per CPU core.
However, this turned out to be complex and wasteful of resources.
Typically, only a small number of runqueues will be active at a
time. Moreover, as kernel threads transition between active and
idle states, it will be much easier to detect when it's safe to
no longer poll a runqueue if it corresponds directly to a thread.

Performance is also improved by this change. For example, the
test_runtime_thread efficiency value increased from 85% to 92%.

The drawback is we're no longer intelligent about hyperthreads. In
a future change, we should tweak the work stealing mechanism to
prefer stealing from the local hyperthread pair first. Also, the
iokernel may need to poll more queues. However, I think this will
be fine as long as we can be smart about only polling active
queues.