runtime: major pass on RCU support
This change splits out the creation and deletion of kthreads to a
new file, laying the ground work for dynamically adding and removing
kthreads at run time.

Using this new infrastructure, it then defines a quiescent period
based on kthread reschedules, so that unreachable RCU objects
can be freed safely.

A number of fixes to the original RCU header are also included as
well as the addition acquire_consume(), a new atomic barrier that
provides the consume semantics required for RCU.